orchestration/requirements.txt
apache-airflow-providers-google==16.1.0
google-cloud-firestore
google-cloud-storage
google-cloud-bigquery